refactor(recommendList) : independent recommendList fetcher from company
